requirements.txt
fastapi
uvicorn
openai
pinecone-client
